refactor. GameRule 패키지 변경(Service->Util)
