+++ src/main/java/step3/utils/GameRule.java
@@ -1,6 +1,4 @@
-package step3.service;
-
-import step3.utils.NumberGenerator;
+package step3.utils;
 
 public class GameRule {
     private static final int STANDARD_FOR_MOVING = 4;
